Rebalance Will on randomized recruits and promotion levels
This rebalance factors in the changes of previous commits that
guarantee that a soldier loses 10-20 Will per mission on turns,
and possibly lose 10-20+ more Will on adverse events each mission.

As such, this balance establishes that 50 will should be the average will
for all endgame units.

At 50 will, a unit can go on the final mission (20-30 turns) and
maintain a low chance to panic (< 30% chance) if wounded. Or a long
mission (20 turn) and return to base unfatigued if few adverse events
occured.

Also at 50 will, a soldier is still vulnerable to Psi attacks from
Sectoids (90 psi), Spectres (100 psi) and Sectoid Commanders (110 psi)
Keeping in mind that in the new system, a soldier who starts at 50 Will
will probably have about 40-45 Will to defend with when they encounter
these enemies mid-mission.
So this does not break psi balance.

Will is granted to units on promotion. On average, 10 Will is added to units.
Specialists & Sharpshooters are designated as
low-will classes, while Assault, Grenadier, and Technical are high-Will.
Will was also added to Psi-Operative and Hero class promotions.

Finally, skill randomization for recruits was rebalanced to make Will
fit in with other skills. The rookie range is now 20-45.
Will no longer detriments or is opposed to PsiOffence in swaps, so
a high Will + high Psi unit should be more common. Will was placed in
greater opposition to hacking and dodge - so high-will but lower hacking
and lower dodge units may appear more.